+++ cspell.json
@@ -1,7 +1,7 @@
 {
     "words": [
         "pyenv",
-        "pyenv_tool",
+        "pyenvtool",
         "Matthew Wyant",
 
         "rgxs",
